Remove raspbi boot files, allow partial deploy

+++ deploy.sh
@@ -5,9 +5,13 @@ if [ -z $1 ] ; then
   exit 1
 fi
 
+if [ "$2" != "" ] ; then
+  tgt="$2"
+else
+  tgt="all"
+fi
 
-mkdir -p dl
-
+# mkdir -p dl
 # if [ ! -f dl/compiler.jar ] ; then
 #   if [ ! -f dl/compiler-latest.zip ] ; then
 #     wget -P dl https://dl.google.com/closure-compiler/compiler-latest.zip
@@ -16,10 +20,15 @@ mkdir -p dl
 #   unzip -d dl dl/compiler-latest.zip
 # fi
 
-rm -f www/combo.js
-# cat www/js/*.js > www/combo-max.js
-# java -jar dl/compiler.jar --third_party --compilation_level=SIMPLE --js=www/combo-max.js --js_output_file=www/combo.js
-cat www/js/*.js > www/combo.js
-scp -r www/* www-data@${1}:/var/www/html/
-scp -r etc/* root@${1}:/etc/
-scp -r boot/* root@${1}:/boot/
+if [ $tgt == "all" ] || [ $tgt == "www" ] ; then
+  rm -f www/combo.js
+  # cat www/js/*.js > www/combo-max.js
+  # java -jar dl/compiler.jar --third_party --compilation_level=SIMPLE --js=www/combo-max.js --js_output_file=www/combo.js
+  cat www/js/*.js > www/combo.js
+  scp -r www/* www-data@${1}:/var/www/html/
+fi
+
+if [ $tgt == "all" ] || [ $tgt == "etc" ] ; then
+  scp -r etc/* root@${1}:/etc/
+fi
+
